Move functions from common to util
